backend/requirements.txt
Django==5.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
black==24.1.1
flake8==7.0.0
isort==5.13.2
